requirements.txt
# Python学习计划 - 依赖包列表

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0

# 文件处理
openpyxl>=3.1.0
pyyaml>=6.0

# 大模型开发（可选）
openai>=1.0.0
anthropic>=0.7.0

# 其他工具
python-dotenv>=1.0.0

